# web/api/v1/testdata/openapi_3.1_golden.yaml
openapi: 3.1.0
info:
    title: Prometheus API
    description: Prometheus is an Open-Source monitoring system with a dimensional data model, flexible query language, efficient time series database and modern alerting approach.
    contact:
        name: Prometheus Community
        url: https://prometheus.io/community/
    version: 0.0.1-undefined
servers:
    - url: /api/v1
paths:
    /query:
        get:
            tags:
                - query
            summary: Evaluate an instant query
            operationId: query
            parameters:
                - name: limit
                  in: query
                  description: The maximum number of metrics to return.
                  required: false
                  explode: false
                  schema:
                    type: integer
                    format: int64
                  examples:
                    example:
                        value: 100
                - name: time
                  in: query
                  description: The evaluation timestamp (optional, defaults to current time).
                  required: false
                  explode: false
                  schema:
                    oneOf:
                        - type: string
                          format: date-time
                          description: RFC3339 timestamp.
                        - type: number
                          format: unixtime
                          description: Unix timestamp in seconds.
                    description: Timestamp in RFC3339 format or Unix timestamp in seconds.
                  examples:
                    RFC3339:
                        value: "2026-01-02T13:37:00Z"
                    epoch:
                        value: 1767361020
                - name: query
                  in: query
                  description: The PromQL query to execute.
                  required: true
                  explode: false
                  schema:
                    type: string
                  examples:
                    example:
                        value: up
                - name: timeout
                  in: query
                  description: Evaluation timeout. Optional. Defaults to and is capped by the value of the -query.timeout flag.
                  required: false
                  explode: false
                  schema:
                    type: string
                  examples:
                    example:
                        value: 30s
                - name: lookback_delta
                  in: query
                  description: Override the lookback period for this query. Optional.
                  required: false
                  explode: false
                  schema:
                    type: string
                  examples:
                    example:
                        value: 5m
                - name: stats
                  in: query
                  description: When provided, include query statistics in the response. The special value 'all' enables more comprehensive statistics.
                  required: false
                  explode: false
                  schema:
                    type: string
                  examples:
                    example:
                        value: all
            responses:
                "200":
                    description: Query executed successfully.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/QueryOutputBody'
                            examples:
                                vectorResult:
                                    summary: 'Instant vector query: up'
                                    value: {"status": "success", "data": {"resultType": "vector", "result": [{"metric": {"__name__": "up", "instance": "demo.prometheus.io:9090", "job": "prometheus"}, "value": [1767436620, "1"]}, {"metric": {"__name__": "up", "env": "demo", "instance": "demo.prometheus.io:9093", "job": "alertmanager"}, "value": [1767436620, "1"]}]}}
                                scalarResult:
                                    summary: 'Scalar query: scalar(42)'
                                    value:
                                        data:
                                            result:
                                                - 1767436620
                                                - "42"
                                            resultType: scalar
                                        status: success
                                matrixResult:
                                    summary: 'Range vector query: up[5m]'
                                    value: {"status": "success", "data": {"resultType": "matrix", "result": [{"metric": {"__name__": "up", "instance": "demo.prometheus.io:9090", "job": "prometheus"}, "values": [[1767436320, "1"], [1767436620, "1"]]}]}}
                default:
                    description: Error executing query.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                tsdbNotReady:
                                    summary: TSDB not ready
                                    value:
                                        error: TSDB not ready
                                        errorType: internal
                                        status: error
        post:
            tags:
                - query
            summary: Evaluate an instant query
            operationId: query-post
            requestBody:
                description: Submit an instant query. This endpoint accepts the same parameters as the GET version.
                content:
                    application/x-www-form-urlencoded:
                        schema:
                            $ref: '#/components/schemas/QueryPostInputBody'
                        examples:
                            simpleQuery:
                                summary: Simple instant query
                                value:
                                    query: up
                            queryWithTime:
                                summary: Query with specific timestamp
                                value:
                                    query: up{job="prometheus"}
                                    time: "2026-01-02T13:37:00.000Z"
                            queryWithLimit:
                                summary: Query with limit and statistics
                                value:
                                    limit: 100
                                    query: rate(prometheus_http_requests_total{handler="/api/v1/query"}[5m])
                                    stats: all
                required: true
            responses:
                "200":
                    description: Instant query executed successfully.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/QueryOutputBody'
                            examples:
                                vectorResult:
                                    summary: 'Instant vector query: up'
                                    value: {"status": "success", "data": {"resultType": "vector", "result": [{"metric": {"__name__": "up", "instance": "demo.prometheus.io:9090", "job": "prometheus"}, "value": [1767436620, "1"]}, {"metric": {"__name__": "up", "env": "demo", "instance": "demo.prometheus.io:9093", "job": "alertmanager"}, "value": [1767436620, "1"]}]}}
                                scalarResult:
                                    summary: 'Scalar query: scalar(42)'
                                    value:
                                        data:
                                            result:
                                                - 1767436620
                                                - "42"
                                            resultType: scalar
                                        status: success
                                matrixResult:
                                    summary: 'Range vector query: up[5m]'
                                    value: {"status": "success", "data": {"resultType": "matrix", "result": [{"metric": {"__name__": "up", "instance": "demo.prometheus.io:9090", "job": "prometheus"}, "values": [[1767436320, "1"], [1767436620, "1"]]}]}}
                default:
                    description: Error executing instant query.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                tsdbNotReady:
                                    summary: TSDB not ready
                                    value:
                                        error: TSDB not ready
                                        errorType: internal
                                        status: error
    /query_range:
        get:
            tags:
                - query
            summary: Evaluate a range query
            operationId: query-range
            parameters:
                - name: limit
                  in: query
                  description: The maximum number of metrics to return.
                  required: false
                  explode: false
                  schema:
                    type: integer
                    format: int64
                  examples:
                    example:
                        value: 100
                - name: start
                  in: query
                  description: The start time of the query.
                  required: true
                  explode: false
                  schema:
                    oneOf:
                        - type: string
                          format: date-time
                          description: RFC3339 timestamp.
                        - type: number
                          format: unixtime
                          description: Unix timestamp in seconds.
                    description: Timestamp in RFC3339 format or Unix timestamp in seconds.
                  examples:
                    RFC3339:
                        value: "2026-01-02T12:37:00Z"
                    epoch:
                        value: 1767357420
                - name: end
                  in: query
                  description: The end time of the query.
                  required: true
                  explode: false
                  schema:
                    oneOf:
                        - type: string
                          format: date-time
                          description: RFC3339 timestamp.
                        - type: number
                          format: unixtime
                          description: Unix timestamp in seconds.
                    description: Timestamp in RFC3339 format or Unix timestamp in seconds.
                  examples:
                    RFC3339:
                        value: "2026-01-02T13:37:00Z"
                    epoch:
                        value: 1767361020
                - name: step
                  in: query
                  description: The step size of the query.
                  required: true
                  explode: false
                  schema:
                    type: string
                  examples:
                    example:
                        value: 15s
                - name: query
                  in: query
                  description: The query to execute.
                  required: true
                  explode: false
                  schema:
                    type: string
                  examples:
                    example:
                        value: rate(prometheus_http_requests_total{handler="/api/v1/query"}[5m])
                - name: timeout
                  in: query
                  description: Evaluation timeout. Optional. Defaults to and is capped by the value of the -query.timeout flag.
                  required: false
                  explode: false
                  schema:
                    type: string
                  examples:
                    example:
                        value: 30s
                - name: lookback_delta
                  in: query
                  description: Override the lookback period for this query. Optional.
                  required: false
                  explode: false
                  schema:
                    type: string
                  examples:
                    example:
                        value: 5m
                - name: stats
                  in: query
                  description: When provided, include query statistics in the response. The special value 'all' enables more comprehensive statistics.
                  required: false
                  explode: false
                  schema:
                    type: string
                  examples:
                    example:
                        value: all
            responses:
                "200":
                    description: Range query executed successfully.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/QueryRangeOutputBody'
                            examples:
                                matrixResult:
                                    summary: 'Range query: rate(prometheus_http_requests_total[5m])'
                                    value: {"status": "success", "data": {"resultType": "matrix", "result": [{"metric": {"__name__": "up", "instance": "demo.prometheus.io:9090", "job": "prometheus"}, "values": [[1767433020, "1"], [1767434820, "1"], [1767436620, "1"]]}]}}
                default:
                    description: Error executing range query.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                tsdbNotReady:
                                    summary: TSDB not ready
                                    value:
                                        error: TSDB not ready
                                        errorType: internal
                                        status: error
        post:
            tags:
                - query
            summary: Evaluate a range query
            operationId: query-range-post
            requestBody:
                description: Submit a range query. This endpoint accepts the same parameters as the GET version.
                content:
                    application/x-www-form-urlencoded:
                        schema:
                            $ref: '#/components/schemas/QueryRangePostInputBody'
                        examples:
                            basicRange:
                                summary: Basic range query
                                value:
                                    end: "2026-01-02T13:37:00.000Z"
                                    query: up
                                    start: "2026-01-02T12:37:00.000Z"
                                    step: 15s
                            rateQuery:
                                summary: Rate calculation over time range
                                value:
                                    end: "2026-01-02T13:37:00.000Z"
                                    query: rate(prometheus_http_requests_total{handler="/api/v1/query"}[5m])
                                    start: "2026-01-02T12:37:00.000Z"
                                    step: 30s
                                    timeout: 30s
                required: true
            responses:
                "200":
                    description: Range query executed successfully.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/QueryRangeOutputBody'
                            examples:
                                matrixResult:
                                    summary: 'Range query: rate(prometheus_http_requests_total[5m])'
                                    value: {"status": "success", "data": {"resultType": "matrix", "result": [{"metric": {"__name__": "up", "instance": "demo.prometheus.io:9090", "job": "prometheus"}, "values": [[1767433020, "1"], [1767434820, "1"], [1767436620, "1"]]}]}}
                default:
                    description: Error executing range query.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                tsdbNotReady:
                                    summary: TSDB not ready
                                    value:
                                        error: TSDB not ready
                                        errorType: internal
                                        status: error
...
    /format_query:
        get:
            tags:
                - query
            summary: Format a PromQL query
            operationId: format-query
            parameters:
                - name: query
                  in: query
                  description: PromQL expression to format.
                  required: true
                  explode: false
                  schema:
                    type: string
                  examples:
                    example:
                        value: sum(rate(http_requests_total[5m])) by (job)
            responses:
                "200":
                    description: Query formatted successfully.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/FormatQueryOutputBody'
                            examples:
                                formattedQuery:
                                    summary: Formatted PromQL query
                                    value:
                                        data: sum by(job, status) (rate(http_requests_total[5m]))
                                        status: success
                default:
                    description: Error formatting query.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                tsdbNotReady:
                                    summary: TSDB not ready
                                    value:
                                        error: TSDB not ready
                                        errorType: internal
                                        status: error
        post:
            tags:
                - query
            summary: Format a PromQL query
            operationId: format-query-post
            requestBody:
                description: Submit a PromQL query to format. This endpoint accepts the same parameters as the GET version.
                content:
                    application/x-www-form-urlencoded:
                        schema:
                            $ref: '#/components/schemas/FormatQueryPostInputBody'
                        examples:
                            simpleFormat:
                                summary: Format a simple query
                                value:
                                    query: up{job="prometheus"}
                            complexFormat:
                                summary: Format a complex query
                                value:
                                    query: sum(rate(http_requests_total[5m])) by (job, status)
                required: true
            responses:
                "200":
                    description: Query formatting completed successfully.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/FormatQueryOutputBody'
                            examples:
                                formattedQuery:
                                    summary: Formatted PromQL query
                                    value:
                                        data: sum by(job, status) (rate(http_requests_total[5m]))
                                        status: success
                default:
                    description: Error formatting query.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                tsdbNotReady:
                                    summary: TSDB not ready
                                    value:
                                        error: TSDB not ready
                                        errorType: internal
                                        status: error
...
    /alertmanagers:
        get:
            tags:
                - alerts
            summary: Get Alertmanager discovery
            operationId: alertmanagers
            responses:
                "200":
                    description: Alertmanager targets retrieved successfully.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AlertmanagersOutputBody'
                            examples:
                                alertmanagerDiscovery:
                                    summary: Alertmanager discovery results
                                    value:
                                        data:
                                            activeAlertmanagers:
                                                - url: http://demo.prometheus.io:9093/api/v2/alerts
                                            droppedAlertmanagers: []
                                        status: success
                default:
                    description: Error retrieving Alertmanager targets.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                tsdbNotReady:
                                    summary: TSDB not ready
                                    value:
                                        error: TSDB not ready
                                        errorType: internal
                                        status: error
...
    /read:
        post:
            tags:
                - remote
            summary: Remote read endpoint
            description: Prometheus remote read endpoint for federated queries. Accepts and returns Protocol Buffer encoded data.
            operationId: remoteRead
            responses:
                "204":
                    description: No Content
                default:
                    description: Error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
...
components:
    schemas:
...
        QueryOutputBody:
            type: object
            properties:
                status:
                    type: string
                    enum:
                        - success
                        - error
                    description: Response status.
                    example: success
                data:
                    $ref: '#/components/schemas/QueryData'
                warnings:
                    type: array
                    items:
                        type: string
                    description: Only set if there were warnings while executing the request. There will still be data in the data field.
                infos:
                    type: array
                    items:
                        type: string
                    description: Only set if there were info-level annotations while executing the request.
            required:
                - status
                - data
            additionalProperties: false
            description: Response body for instant query.
...